requirements.txt
numpy
pandas
python-binance
requests
ta
schedule
python-dotenv
termcolor
python-telegram-bot